+++ Peace-Treaty-V1.2/Operations/Source-Files/Misc/Input.cpp
@@ -113,7 +113,7 @@ std::string selectOption(std::vector<std::string>AV) {
 	std::cout << "Enter an option: " << LOG::getColor(LOG::BLUE);  
 	std::string input;
 	getline(std::cin, input);
-	for (int i = 0; i < input.size(); i++) {
+	for (size_t i = 0; i < input.size(); i++) {
 		input[i] = toupper(input[i]);
 	}
 	LOG::addColor(LOG::RESET);
