+++ Peace-Treaty-V1.2/Operations/Source-Files/Units/PrimeUnits/PrimeUnits.cpp
@@ -147,7 +147,7 @@ unitSPTRList PrimeUnits::levelSort(unitSPTRList list) {
 		return list;
 	}
 
-	int listSize = list.size();
+	int listSize = (int)list.size();
 	unitSPTR lastElement = list.at(listSize - 1);
 	int lastLevel = lastElement->getLevel();
